Adds new rake tasks and abstracts URI
This request adds three new rake tasks:

 - clean: to delete the tmp directory
 - delete: to delete the Solr index
 - all: combines clone and index for easier first time use

Also, abstracts the Solr URL for better handling of alternative Solr
installations. Use of an .env file is supported, as well as adding the
URL to the rake tasks (index, delete, and all).

Additionally, the --verbose flag can be used to see output.