Ensure that locn_geometry is populated when migrating v1 to Aardvark
Previously we were setting dcat_bbox but not locn_geometry based
on the v1 solr_geom. The latter is also crucial, because we can't
do spatial search or render previews without it.

This ensures that we always populate locn_geometry if we had solr_geom,
and then also fill in dcat_bbox if the geometry happens to be in the
ENVELOPE() format.